% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pipe_source}
\alias{get_pipe_source}
\title{Get source argument from a chain}
\usage{
get_pipe_source(x)
}
\arguments{
\item{x}{magrittr}
}
\description{
Magrittr pipe function
}
\details{
`lazyeval::expr_text` does not work for pipes ...
}
\examples{

 f <- . \%>\% mean( na.rm = TRUE )
 nm_mean <-  mean \%>\% set_namer()

 get_pipe_source(f)
}
\seealso{
- `lazyeval::expr_text`
}
