% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{name}
\alias{name}
\alias{name.train.formula}
\title{name}
\usage{
name(x, ...)

\method{name}{train.formula}(x, sep = ".", ...)
}
\arguments{
\item{x}{object}

\item{...}{}
}
\description{
Provide a canonical name for an object
}
\details{
\code{name} provides a canonical name for an object based on it's content and
attributes
}
\examples{

  name(1:10)
  name(letters)
  name(iris)
  
}
