% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_it.R
\name{name_it}
\alias{name_it}
\alias{name_it.character}
\alias{name_it.default}
\title{Automatically name metrics}
\usage{
name_it(x, fun)

\method{name_it}{character}(x, fun)

\method{name_it}{default}(x, fun)
}
\arguments{
\item{fun}{`with_naming` function}

\item{base}{string that serves a the root to the name and passed as the first
argument to the namer function}
}
\description{
Build a name from a naming function
}
\details{
A name is comprised of a prefix, base and suffix
}
\examples{

nm_mean <- mean \%>\% set_namer

"price" \%>\% name_it( nm_mean )

 retail <- rnorm(50)
 
 retail \%>\% name_it(nm_mean)


}
\references{
- Cf. Kurremkarmerruk (the master namer)
}
