% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_prefix.R
\name{nm_prefix}
\alias{nm_prefix}
\alias{nm_prefix.default}
\alias{nm_prefix.data.frame}
\alias{nm_prefix.data.table}
\alias{nm_suffix.default}
\alias{nm_suffix.data.frame}
\alias{nm_suffix.data.table}
\title{Add a prefix or suffix to the names of an object}
\usage{
nm_prefix(x, prefix, sep = "", ...)

\method{nm_prefix}{default}(x, prefix, sep = "", ...)

\method{nm_prefix}{data.frame}(x, prefix, sep = "", ...)

\method{nm_prefix}{data.table}(x, prefix, sep = "", exclude_key = TRUE)

\method{nm_suffix}{default}(x, suffix, sep = "", ...)

\method{nm_suffix}{data.frame}(x, suffix, sep = "", ...)

\method{nm_suffix}{data.table}(x, suffix, sep = "", ..., exclude_key)
}
\arguments{
\item{x}{object with a names attribute}

\item{prefix}{character; prefix collapsed to string by \code{sep} and
added to the front of each name}

\item{sep}{string used to collapse and seperate \code{prefix} or
\code{suffix} from names. The default is not to use a separator.}

\item{...}{additional arguments}

\item{exclude_key}{logical; if \code{TRUE} do not mangle the object's keys.
(DEFAULT: \code{TRUE})}

\item{suffix}{character; collapsed to string by \code{sep} and
added to the back of each name.}
}
\value{
An object with the names altered by a prefix or suffix. For data.tables,
these are returned invisibly.
}
\description{
Adds a prefix or suffix to the names attributes.
}
\details{
\code{nm_prefix} and \code{nm_suffix} add a prefix or suffix to the names of
an object.

The method for data.table objects are modified by reference.
}
\examples{

# nm_prefix

  data(iris)
  iris <- nm_prefix( iris, "var_" )
  names(iris)

  data(iris)
  setDT(iris)
  nm_prefix( iris, "var_" )
  names(iris)

# nm_suffix
  data(iris)
  iris <- nm_suffix( iris, "_var" )
  names(iris)

  data(iris)
  setDT(iris)
  nm_suffix( iris, "_var" )
  names(iris)

}
\seealso{
\code{\link[stringr.tools:str_prefix]{stringr.tools::str_prefix()}} which is used by these functions

(setnames)\code{\link[data.table:setattr]{data.table::setattr()}}

(data.table::setnames)\code{\link[data.table:setattr]{data.table::setattr()}}
}
