% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_prefix_if.R
\name{nm_prefix_if}
\alias{nm_prefix_if}
\alias{nm_prefix_if.default}
\alias{nm_prefix_if.data.table}
\alias{nm_prefix_at}
\alias{nm_prefix_at.data.table}
\title{Conditionally prefix names}
\usage{
nm_prefix_if(.x, .predicate, .prefix)

\method{nm_prefix_if}{default}(.x, .predicate, .prefix)

\method{nm_prefix_if}{data.table}(.x, .predicate, .prefix)

nm_prefix_at(.x, .vars, .prefix)

nm_prefix_at(.x, .vars, .prefix)

nm_prefix_at.data.table(.x, .vars, .prefix)
}
\arguments{
\item{.x}{object containing named elements}

\item{.predicate}{function applied to each element of \code{.x} yielding a logical
vector indicating which names to prefix.}

\item{.prefix}{string to be used as the prefix}
}
\description{
Conditionally prefix names
}
\examples{
  data(iris)
  setDT(iris)
  nm_prefix_if( iris, is.factor,  "prefix." )

  data(iris)
  nm_prefix_at( iris, 'Species', 'pre.' )

}
