% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_replace.R
\name{nm_replace}
\alias{nm_replace}
\title{Replaced matched patterns in names}
\usage{
nm_replace(x, pattern, replacement)
}
\arguments{
\item{x}{object with names}

\item{pattern}{pattern to match}

\item{replacement}{replacement vector}
}
\description{
Replaced matched patterns in names
}
\details{
\code{nm_replace} is like \code{str_replace} but operates on the \code{names} attribute
of \code{x} rather than a user-supplied string.
}
\examples{

v <- c( aa=1, ab=2, ac=3 )
v

v \%>\% nm_replace( "^a", "b" )

}
