% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.with_naming}
\alias{print.with_naming}
\title{print naming function}
\usage{
\method{print}{with_naming}(...)
}
\arguments{
\item{...}{arguments passed to print}
}
\description{
print naming function
}
