% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print0.R
\name{print0}
\alias{print0}
\title{print0}
\usage{
print0(x, attribs = FALSE, ...)
}
\arguments{
\item{x}{object}

\item{attribs}{logical; whether to print attributes. (Default: \code{FALSE})}

\item{...}{additional arguments to the print method}
}
\description{
Print and object, optionally with attributes
}
\details{
Print an object without its attributes.

It is easy to use this quickly for a new class.  Simply:\preformatted{print.myclass <- base.tools::print0
}
}
\examples{

x <- 1:5 
attr( x, "foo" ) <- "bar"

print(x)
print0(x)
print0( x, attribs = TRUE )

}
\seealso{
\itemize{
\item \code{strip_attributes}
}
}
