% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_name_after.R
\name{the_name_after}
\alias{the_name_after}
\title{the_name_after}
\usage{
the_name_after(.data, name)
}
\arguments{
\item{.data}{object; an object with a names vector, usually a table}

\item{name}{}
}
\value{
string; the name
}
\description{
Get the name following the name provided
}
\details{
`the_name_after` returnes
}
