% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_namer.R
\name{set_namer}
\alias{set_namer}
\alias{set_namer.default}
\alias{set_namer.function}
\alias{set_namer.fseq}
\alias{get_namer}
\title{set_namer / get_namer}
\usage{
set_namer(fun, namer, ...)

\method{set_namer}{default}(fun, namer, ...)

\method{set_namer}{function}(fun, namer = NULL, ...)

\method{set_namer}{fseq}(fun, namer = NULL, ...)

get_namer(fun)
}
\arguments{
\item{fun}{function or fseq to convert to a naming function}

\item{namer}{function or string; See Details.}
}
\value{
\code{fun} with appended class 'with_namer' and the \code{namer} attribute set.
}
\description{
sets or gets the \code{namer} attribute of a function
}
\details{
\code{set_namer} is a function decorator that adds/updates a \code{namer`` attribute of }fun`. It is also a mixin/role for adding additional capabilities to a
function.

The \code{namer} argument can either be a string or function. If a function, this
is used as the naming function.  string, the
naming function is to use this function as a suffix    `

\code{get_namer} retrieves the naming function.
}
\examples{

# With functions:

 f1 <- mean \%>\% set_namer( function(x) str_suffix(x,'.mean') )
 "price" \%>\% name_it(f1)       # "price.mean"

 f2 <- mean \%>\% set_namer( . \%>\% str_suffix('.mean') )
 "price" \%>\% name_it(f2)       # "price.mean"

 f3 <- mean \%>\% set_namer('.mean')
 "price" \%>\% name_it(f3)       # "price.mean"

 f4 <- mean \%>\% set_namer()
 "price" \%>\% name_it(f4)       # "price.mean"

 f5 <- set_namer(mean)
 "price" \%>\% name_it(f5)       # "price.mean"

# With fseq:
 f <- . \%>\% mean( na.rm=TRUE )
 f <- set_namer(f)
 "price" \%>\% name_it(f)        # "price.mean"


}
\seealso{
\itemize{
\item `name_it``
}
}
